refactor(controllers): own resume refusal presentation
